Create a new cache if we upgrade to a new version of Gradle.
Otherwise we'll re-download Gradle on every build. (Caches are immutable
in Circle CI, so we would never update the cache with the new Gradle
version.)

Every time we upgrade Gradle, we have to change gradle.properties (to
include the new distribution URL), so adding a checksum of that file
will force CircleCI to create a new cache when we upgrade.